+++ noxfile.py
@@ -1,7 +1,19 @@
 import nox
 
+def confirm(message, validators = ["y", "yes"]):
+    valid_input = False
+
+    while not valid_input:
+        response = input(message + "(")
+        if validators in response:
+            valid_input = True
+
 @nox.session
 def build(session):
+    confirm("Have you run the tests?")
+    confirm("Have you updated inline docs?")
+    confirm("Have you updated the wiki docs?")
+    confirm("Have you updated the readme docs?")
     session.run("python", "setup.py", "sdist")
     session.install("wheel")
     session.run("python", "setup.py", "bdist_wheel", "--universal")