+++ tests/primary_test.py
@@ -1,0 +1,14 @@
+def test_library():
+    """Validates that the resource library works"""
+
+def test_executables():
+    """Validates that the varying forms of executables work"""
+
+def test_packages():
+    """Validates that packages installed via package manager work"""
+
+def test_archives():
+    """Validates that the varying forms of archive extraction work"""
+
+def test_static():
+    """Validates that downloading of static assets works"""